Extract frontend types to new files
