+++ client/src/lib/types/encryptedMessagesHub.ts
@@ -1,0 +1,6 @@
+﻿import type { EncryptedDirectMessage } from "./index";
+
+export type EncryptedDirectChatUpdatedPayload = {
+  encryptedDirectChatId?: string;
+  message?: EncryptedDirectMessage;
+};
