+++ client/src/lib/types/fileUpload.ts
@@ -1,0 +1,26 @@
+﻿import type { MediaUploadResult, MessageType } from "./index";
+
+export type SelectedItem = {
+  id: string;
+  file: File;
+  kind: "image" | "video" | "audio" | "other";
+  preview?: string | null;
+};
+
+export interface FileUploadState {
+  selectedItems: SelectedItem[];
+  pendingPaste: {
+    file: File | null;
+    preview: string | null;
+  };
+}
+
+export interface UseFileUploadProps {
+  chatRoomId?: string;
+  onUpload: (
+    body: string,
+    type: MessageType,
+    mediaData?: Partial<MediaUploadResult>
+  ) => Promise<void>;
+  onReset: () => void;
+}
