+++ client/src/lib/types/hubHandlers.ts
@@ -1,0 +1,4 @@
+﻿export type RegisteredHandler = {
+  event: string;
+  handler: (...args: unknown[]) => void;
+};
