+++ client/src/lib/types/messagesHook.ts
@@ -1,0 +1,5 @@
+﻿import type { ChatMessage, PagedList } from "./index";
+
+export type MessagesPayload =
+  | { chatRoomId?: string; data?: PagedList<ChatMessage, Date> }
+  | PagedList<ChatMessage, Date>;
