+++ client/src/lib/types/scrollHandler.ts
@@ -1,0 +1,5 @@
+﻿import type { BaseMessageStore } from "./index";
+
+export interface UseScrollHandlerProps {
+  messageStore: BaseMessageStore;
+}
